+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,9 @@
         android:theme="@style/Base.Theme.VaccinationApp"
         tools:targetApi="31">
         <activity
+            android:name=".HashClass"
+            android:exported="false" />
+        <activity
             android:name=".VaccineControl.testLogin"
             android:exported="false" />
         <activity
