+++ app/src/main/java/com/example/vaccinationapp/HashClass.kt
@@ -1,0 +1,19 @@
+package com.example.vaccinationapp
+
+import androidx.appcompat.app.AppCompatActivity
+import java.security.MessageDigest
+
+open class HashClass : AppCompatActivity() {
+
+    fun hashData(data: String): String {
+        val bytes = data.toByteArray()
+        val md = MessageDigest.getInstance("SHA-256")
+        val digest = md.digest(bytes)
+        return digest.toHexString()
+    }
+
+    fun ByteArray.toHexString(): String {
+        return joinToString("") { "%02x".format(it) }
+    }
+
+}
