+++ app/src/main/java/com/example/vaccinationapp/SignUpActivity.kt
@@ -1,7 +1,6 @@
 package com.example.vaccinationapp
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.text.TextUtils
 import android.util.Patterns
@@ -17,12 +16,11 @@ import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
-import java.security.MessageDigest
 
 import java.util.UUID
 
 
-open class SignUpActivity : AppCompatActivity() {
+open class SignUpActivity : HashClass() {
 
     private lateinit var goToSignInButton: Button
     private lateinit var registerEmail: EditText
@@ -45,15 +43,14 @@ open class SignUpActivity : AppCompatActivity() {
         goToSignIn()
         registerNewUserButton()
 
+
     }
 
     private fun registerNewUserButton() {
         registerButton.setOnClickListener {
             addUserToPhp()
-
             val intent = Intent(this, SignInActivity::class.java)
             startActivity(intent)
-
         }
     }
 
@@ -62,11 +59,6 @@ open class SignUpActivity : AppCompatActivity() {
             val intent = Intent(this, SignInActivity::class.java)
             startActivity(intent)
         }
-    }
-
-    private fun validateEmail(email: String): Boolean {
-        val emailPatter = Patterns.EMAIL_ADDRESS
-        return emailPatter.matcher(email).matches()
     }
 
     private fun registrationData(): Boolean {
@@ -106,45 +98,6 @@ open class SignUpActivity : AppCompatActivity() {
         }
     }
 
-    //    private fun registerUser() {
-//        // email used for registration
-//        val email: String = registerEmail.text.toString().trim(' ')
-//
-//        // password chosen in registration
-//        val password: String = registerPassword.text.toString().trim(' ')
-//
-//        // name according to registered Name
-//        val name: String = registerName.text.toString().trim(' ')
-//
-//        if (registrationData()) {
-//            FirebaseAuth.getInstance().createUserWithEmailAndPassword(email, password)
-//                .addOnCompleteListener(
-//                    OnCompleteListener<AuthResult> { task ->
-//                        if (task.isSuccessful) {
-//                            val firebaseUser: FirebaseUser = task.result!!.user!!
-//                            makeToast("Registration successful, user Id ${firebaseUser.uid}", false)
-//
-//                            val user = User(
-//                                "ID",
-//                                name,
-//                                true,
-//                                email,
-//                            )
-//
-//                            FireStoreClass().registerUserFS(this@SignUpActivity, user)
-//                            finish()
-//
-//                        } else {
-//                            Toast.makeText(
-//                                this@SignUpActivity, resources.getString(R.string.register_failed),
-//                                Toast.LENGTH_LONG
-//                            ).show()
-//                        }
-//                    }
-//                )
-//        }
-//
-//    }
 
     @OptIn(DelicateCoroutinesApi::class)
     private fun addUserToPhp() {
@@ -152,82 +105,43 @@ open class SignUpActivity : AppCompatActivity() {
         val password: String = registerPassword.text.toString().trim()
         val userId = generateUserId()
 
-        // Launch a coroutine to perform database operations asynchronously
-        GlobalScope.launch(Dispatchers.IO) {
-            try {
-                val connection = DBConnection.getConnection()
-                val dbQueries = DBQueries(connection)
+        if (registrationData()) {
+            // Launch a coroutine to perform database operations asynchronously
+            GlobalScope.launch(Dispatchers.IO) {
+                try {
+                    val connection = DBConnection.getConnection()
+                    val dbQueries = DBQueries(connection)
 
-                val newUser = SignUpDataClass(
-                    userId,
-                    hashData(email),
-                    hashData(password)
-                )
+                    val newUser = SignUpDataClass(
+                        userId,
+                        hashData(email),
+                        hashData(password)
+                    )
+                    val isSuccessful = dbQueries.insertUser(newUser)
+                    if (isSuccessful) {
+                        userRegistrationSuccess()
+                        connection.close()
+                    } else {
+                        makeToast("Unable to Add User", false)
+                    }
 
-                val isSuccessful = dbQueries.insertUser(newUser)
-
-                if (isSuccessful) {
-                    makeToast("Registration Successful userId: $userId ", false)
-                    connection.close()
-                } else {
-                    makeToast("Unable to Add User", false)
+                } catch (e: Exception) {
+                    e.printStackTrace()
                 }
-
-            } catch (e: Exception) {
-                e.printStackTrace()
             }
         }
     }
 
-    //
-//private fun addUserToPhp() {
-    //        val email: String = registerEmail.text.toString().trim()
-//        val password: String = registerPassword.text.toString().trim()
-//
-//        if (registrationData()) {
-//            makeToast("we are here", false)
-//            // Generate user ID
-//            val userId = generateUserId()
-//
-//            // Assuming DBConnection.getConnection() returns a valid database connection
-//            val connection = DBConnection.getConnection()
-//            val dbQueries = DBQueries(connection)
-//            val newUser = LogInDataClass(userId, email, password)
-//
-//            // Insert user data into the database
-//            val isSuccess = dbQueries.insertUser(newUser)
-//            connection.close()
-//
-//            if (isSuccess) {
-//                // User registration successful
-//                makeToast("Registration successful, user Id $userId", false)
-//            } else {
-//                // User registration failed
-//                Toast.makeText(
-//                    this@SignUpActivity,
-//                    resources.getString(R.string.register_failed),
-//                    Toast.LENGTH_LONG
-//                ).show()
-//            }
-//        }
-//    }
+    private fun validateEmail(email: String): Boolean {
+        val emailPatter = Patterns.EMAIL_ADDRESS
+        return emailPatter.matcher(email).matches()
+    }
+
     private fun generateUserId(): String {
         return UUID.randomUUID().toString()
     }
 
-    private fun hashData(data: String): String {
-        val bytes = data.toByteArray()
-        val md = MessageDigest.getInstance("SHA-256")
-        val digest = md.digest(bytes)
-        return digest.toHexString()
-    }
-
-    fun ByteArray.toHexString(): String {
-        return joinToString("") { "%02x".format(it) }
-    }
-
-
-    fun userRegistrationSuccess() {
+    private fun userRegistrationSuccess() {
         Toast.makeText(
             this@SignUpActivity, resources.getString(R.string.register_success),
             Toast.LENGTH_LONG
