+++ app/src/main/java/com/example/vaccinationapp/phpAdmin/DbUser.kt
@@ -2,8 +2,6 @@ package com.example.vaccinationapp.phpAdmin
 
 interface DbUser {
 
-    fun getPassword(password: String): SignUpDataClass?
-   // fun getUser(email: String, password: String): LogInDataClass?
     fun insertUser(user: SignUpDataClass): Boolean
     fun userExists(email: String, password: String): Boolean
